requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5
Pillow>=8.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.0
python-dotenv>=0.19.0
openai>=1.0.0
pydantic>=2.0.0
jinja2>=3.0.0
byaldi>=0.1.0
transformers>=4.0.0
torch>=2.0.0
groq>=0.1.0
ollama>=0.1.0
PyMuPDF>=1.18.0